Server and Client Handling Server include Error Message Handle Error on client side Java Bean Validatio Enforcing Database Constraints Add new student button when no data avalible

+++ src/main/java/com/tilmeez/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.tilmeez.demo.student;
 
+import com.tilmeez.demo.student.exception.BadRequestException;
+import com.tilmeez.demo.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,12 +19,22 @@ public class StudentService {
 
     public void addStudent(Student student) {
         // check if email is taken
-
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " is taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
         // check if student exists
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id: " + studentId + " does not exists"
+            );
+        }
         studentRepository.deleteById(studentId);
     }
 }
